# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Masafah API Documentation",
        "description": "Api description...",
        "version": "1.0.0"
    },
    "paths": {
...
        "/~tvavisa/masafah/public/api/user/verifyOTP": {
            "post": {
                "tags": [
                    "User OTP"
                ],
                "summary": "Verify User OTP",
                "operationId": "verifyOTP",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Verify OTP Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "mobile": {
                                    "description": "user's mobile number",
                                    "type": "integer",
                                    "example": 88663456
                                },
                                "otp": {
                                    "description": "generated OTP",
                                    "type": "integer",
                                    "example": 46103
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "404": {
                        "description": "Mobile not found"
                    },
                    "417": {
                        "description": "Wrong OTP"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/resendOTP": {
            "post": {
                "tags": [
                    "User OTP"
                ],
                "summary": "Resend User OTP",
                "operationId": "resendOTP",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Mobile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "mobile": {
                                    "description": "users mobile number",
                                    "type": "integer",
                                    "example": 88663456
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
...
    },
...
}